# backend/docs/swagger.yaml
definitions:
  auth.AuthResponse:
    properties:
      accessToken:
        type: string
      email:
        type: string
      refreshToken:
        type: string
      status:
        type: integer
    type: object
  types.AddFormUser:
    properties:
      email:
        type: string
      isAdmin:
        type: boolean
      nickname:
        type: string
      password:
        type: string
    required:
    - email
    - nickname
    type: object
  types.LoginForm:
    properties:
      nickname:
        type: string
      password:
        type: string
    required:
    - nickname
    - password
    type: object
  types.RegisterForm:
    properties:
      email:
        type: string
      isAdmin:
        type: boolean
      nickname:
        type: string
      password:
        type: string
    required:
    - email
    - nickname
    - password
    type: object
  types.UpdateFormUser:
    properties:
      email:
        type: string
      id:
        type: string
      isAdmin:
        type: boolean
      nickname:
        type: string
      password:
        type: string
    required:
    - email
    - id
    - nickname
    type: object
  user.User:
    properties:
      email:
        type: string
      id:
        type: integer
      isAdmin:
        type: boolean
      nickname:
        type: string
    type: object
info:
  contact: {}
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Вход по email и паролю
      parameters:
      - description: Данные входа
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/types.LoginForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.AuthResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Авторизация
      tags:
      - auth
  /auth/logout:
    post:
      description: Инвалидация токена
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Выход
      tags:
      - auth
  /auth/me:
    post:
      description: Возвращает текущего пользователя из токена
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.AuthResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Получить информацию о себе
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Получить новый access токен
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.AuthResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Обновление токена
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Создание нового пользователя
      parameters:
      - description: Данные регистрации
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/types.RegisterForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.AuthResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Регистрация
      tags:
      - auth
  /user/add:
    post:
      consumes:
      - application/json
      description: Админ может создать нового пользователя
      parameters:
      - description: Пользователь для создания
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/types.AddFormUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Добавить пользователя
      tags:
      - user
  /user/delete:
    delete:
      description: Админ может удалить любого, пользователь — только себя
      parameters:
      - description: ID пользователя для удаления
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Удалить пользователя
      tags:
      - user
  /user/get:
    get:
      description: Админ может указать nickname в query, обычный пользователь получает
        себя по токену
      parameters:
      - description: Никнейм пользователя (только для админов)
        in: query
        name: nickname
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Получить информацию о пользователе
      tags:
      - user
  /user/update:
    put:
      consumes:
      - application/json
      description: Админ может обновить любого, обычный — только себя
      parameters:
      - description: Данные для обновления
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/types.UpdateFormUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Обновить пользователя
      tags:
      - user
swagger: "2.0"