# backend/pkg/types/auth.go
package types

type LoginForm struct {
	Nickname string `json:"nickname" binding:"required"`
	Password string `json:"password" binding:"required"`
}

type RegisterForm struct {
	Email    string `json:"email" binding:"required,email"`
	Nickname string `json:"nickname" binding:"required"`
	Password string `json:"password" binding:"required, password_strength"`
	IsAdmin  bool   `json:"isAdmin"`
}